chore: move files from pages folder to app folder
