+++ web/src/providers/routing-provider.tsx
@@ -1,11 +1,11 @@
+import { CreatePost } from "@/app/create-post";
+import { Home } from "@/app/home";
+import { Login } from "@/app/login";
+import { NotFound } from "@/app/not-found";
+import { PostDetail } from "@/app/post-detail";
+import { Register } from "@/app/register";
+import { YourPosts } from "@/app/your-posts";
 import { AppLayout } from "@/components/app-layout";
-import { CreatePost } from "@/pages/create-post";
-import { Home } from "@/pages/home";
-import { Login } from "@/pages/login";
-import { NotFound } from "@/pages/not-found";
-import { PostDetail } from "@/pages/post-detail";
-import { Register } from "@/pages/register";
-import { YourPosts } from "@/pages/your-posts";
 import { useAppSelector } from "@/store";
 import {
   Navigate,
